+++ navground_core_py/include/navground/core_py/property.h
@@ -7,6 +7,7 @@
 #include "navground/core_py/yaml.h"
 
 namespace py = pybind11;
+using navground::core::get_scalar_type_name;
 using navground::core::HasProperties;
 using navground::core::Property;
 using navground::core::Vector2;
@@ -17,6 +18,9 @@ template <typename T, typename A>
 struct is_std_vector<std::vector<T, A>> : std::true_type {};
 
 template <class T> constexpr bool is_std_vector_v = is_std_vector<T>::value;
+
+using Converter =
+    std::function<std::optional<Property::Field>(const Property::Field &)>;
 
 bool is_empty_vector(const Property::Field &value) {
   return std::visit(
@@ -81,32 +85,31 @@ std::optional<Property::Field> convert_py(const Property::Field &value) {
 // Pay attention that Python may pass
 // a value like ``[3.2, 1, 4] as [True, True, True]
 
-inline std::optional<Property::Field>
-convert_py(const Property::Field &value, const std::string &scalar_type_name,
-           bool is_vector) {
+inline std::optional<Converter>
+get_converter(const std::string &scalar_type_name, bool is_vector) {
   if (is_vector) {
     if (scalar_type_name == "int")
-      return convert_py<std::vector<int>>(value);
+      return convert_py<std::vector<int>>;
     if (scalar_type_name == "float")
-      return convert_py<std::vector<ng_float_t>>(value);
+      return convert_py<std::vector<ng_float_t>>;
     if (scalar_type_name == "str")
-      return convert_py<std::vector<std::string>>(value);
+      return convert_py<std::vector<std::string>>;
     if (scalar_type_name == "bool")
-      return convert_py<std::vector<bool>>(value);
+      return convert_py<std::vector<bool>>;
     if (scalar_type_name == "vector")
-      return convert_py<std::vector<Vector2>>(value);
+      return convert_py<std::vector<Vector2>>;
     return std::nullopt;
   }
   if (scalar_type_name == "int")
-    return convert_py<int>(value);
+    return convert_py<int>;
   if (scalar_type_name == "float")
-    return convert_py<ng_float_t>(value);
+    return convert_py<ng_float_t>;
   if (scalar_type_name == "str")
-    return convert_py<std::string>(value);
+    return convert_py<std::string>;
   if (scalar_type_name == "bool")
-    return convert_py<bool>(value);
+    return convert_py<bool>;
   if (scalar_type_name == "vector")
-    return convert_py<Vector2>(value);
+    return convert_py<Vector2>;
   return std::nullopt;
 }
 
@@ -123,17 +126,17 @@ convert_py(const Property::Field &value, const Property::Field &default_value) {
 inline Property
 make_property_py(const py::object &py_getter, const py::object &py_setter,
                  const Property::Field &default_value,
+                 const Converter &converter,
                  const std::string &description = "",
                  const std::optional<py::function> &py_schema = std::nullopt,
                  const std::vector<std::string> &deprecated_names = {}) {
   Property p;
   if (!py_getter.is_none()) {
-    p.getter = [py_getter, default_value](const HasProperties *obj) {
+    p.getter = [py_getter, converter, default_value](const HasProperties *obj) {
       // CHANGED(28/5/2025): Added layer of safety here because we
       // are not wrapping python getters, which they may return arbitrary types.
       auto py_value = py_getter.attr("__call__")(py::cast(obj));
-      if (auto v =
-              convert_py(py_value.cast<Property::Field>(), default_value)) {
+      if (auto v = converter(py_value.cast<Property::Field>())) {
         return *v;
       }
       return default_value;
@@ -175,23 +178,59 @@ make_property_py(const py::object &py_getter, const py::object &py_setter,
 
 inline Property make_property_with_py_property(
     const py::object &py_property, const Property::Field &default_value,
-    const std::string &description = "",
+    const Converter &converter, const std::string &description = "",
     const std::optional<py::function> &py_schema = std::nullopt,
     const std::vector<std::string> &deprecated_names = {}) {
   return make_property_py(py_property.attr("fget"), py_property.attr("fset"),
-                          default_value, description, py_schema,
+                          default_value, converter, description, py_schema,
                           deprecated_names);
 }
 
+inline Property make_property_py_with_type(
+    const py::object &py_getter, const py::object &py_setter,
+    const Property::Field &default_value, const std::string &type_name,
+    const std::string &description = "",
+    const std::optional<py::function> &py_schema = std::nullopt,
+    const std::vector<std::string> &deprecated_names = {}) {
+  const auto [scalar_type, is_vector] = get_scalar_type_name(type_name);
+  const auto converter = get_converter(scalar_type, is_vector);
+  if (!converter) {
+    throw std::runtime_error("Invalid type name" + type_name);
+  }
+  const auto safe_default_value = (*converter)(default_value);
+  if (!safe_default_value) {
+    throw std::runtime_error("Invalid default value");
+  }
+  return make_property_py(py_getter, py_setter, *safe_default_value, *converter,
+                          description, py_schema, deprecated_names);
+}
+
+inline Property make_property_with_py_property_with_type(
+    const py::object &py_property, const Property::Field &default_value,
+    const std::string &type_name, const std::string &description = "",
+    const std::optional<py::function> &py_schema = std::nullopt,
+    const std::vector<std::string> &deprecated_names = {}) {
+  const auto [scalar_type, is_vector] = get_scalar_type_name(type_name);
+  const auto converter = get_converter(scalar_type, is_vector);
+  if (!converter) {
+    throw std::runtime_error("Invalid type name" + type_name);
+  }
+  const auto safe_default_value = (*converter)(default_value);
+  if (!safe_default_value) {
+    throw std::runtime_error("Invalid default value");
+  }
+  return make_property_py(py_property.attr("fget"), py_property.attr("fset"),
+                          *safe_default_value, *converter, description,
+                          py_schema, deprecated_names);
+}
+
 inline py::cpp_function safe_setter(py::function setter,
-                                    const std::string &scalar_type,
-                                    bool is_vector) {
+                                    const Converter &converter) {
   return py::cpp_function(
-      [setter, scalar_type, is_vector](py::object self, py::object value) {
-        auto cpp_value = value.cast<Property::Field>();
-        auto typed_cpp_value = convert_py(cpp_value, scalar_type, is_vector);
-        if (typed_cpp_value) {
-          setter(self, py::cast(*typed_cpp_value));
+      [setter, converter](py::object self, py::object value) {
+        auto safe_value = converter(value.cast<Property::Field>());
+        if (safe_value) {
+          setter(self, py::cast(*safe_value));
         }
       });
 }